pyproject.toml
[project]
name = "contributions-importer-for-github"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "gitpython>=3.1.45",
]
